Set up local terraform mechanism for developing POC infrastructure locally
I think it's really important to be able to do development and learning outside of Atlantis,
and this seems like the best way for a dev to spin up infrastructure that can be played with at will.
Because the heavy lifting is done by the cloud functions tf module,
there's not much code here, and it's not too much work to maintain.
There's some copy/paste of locals.tf between here and cru-terraform, but I think it'll be worth it.