Don't depend on github.event.inputs.function_name for function_name input
It looks like it is not populated when a pull request is merged;
it was blank for
https://github.com/CruGlobal/dot/actions/runs/11629510049/job/32386616681\#step:5:1
in particular.

Also, I don't think we need this to be parameterized.
Devs don't have a use case for changing this for manual deployments.
And same for platform and entrypoint; these are all static config.